+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ## Add
 - Add [extend_aspect_ratio](https://docs.imgproxy.net/latest/generating_the_url?id=extend-aspect-ratio) processing option.
+- (pro) Add [advanced smart crop](https://docs.imgproxy.net/latest/configuration?id=smart-crop)
 
 ## [3.13.2] - 2023-02-15
 ### Change