+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [Unreleased]
 ## Add
 - Add [extend_aspect_ratio](https://docs.imgproxy.net/latest/generating_the_url?id=extend-aspect-ratio) processing option.
-- (pro) Add [advanced smart crop](https://docs.imgproxy.net/latest/configuration?id=smart-crop)
+- Add the `IMGPROXY_ALLOW_SECURITY_OPTIONS` config + `max_src_resolution`, `max_src_file_size`, `max_animation_frames`, and `max_animation_frame_resolution` processing options.
+- (pro) Add [advanced smart crop](https://docs.imgproxy.net/latest/configuration?id=smart-crop).
 
 ### Change
 - Make the `expires` processing option set `Expires` and `Cache-Control` headers.